+++ src/main/java/ch/bfh/bti7535/w2016/algorithm/AbstractAlgorithm.java
@@ -1,0 +1,8 @@
+package ch.bfh.bti7535.w2016.algorithm;
+
+import java.util.List;
+
+public abstract class AbstractAlgorithm {
+
+	public abstract List<Document> execute(List<Document> input);
+}
