+++ src/main/java/ch/bfh/bti7535/w2016/algorithm/NaiveBayesAlgorithm.java
@@ -1,0 +1,11 @@
+package ch.bfh.bti7535.w2016.algorithm;
+
+import java.util.List;
+
+public class NaiveBayesAlgorithm extends AbstractAlgorithm {
+
+	@Override
+	public List<Document> execute(List<Document> input) {
+		return null;
+	}
+}
